Split generate_compositions.py into two files, one for breath samples, the other for comps for simulations.
